asm_util::tsc

Struct Tsc

#[repr(C)]
pub struct Tsc {
    pub cyclesPerSecond: u64,
}

Emulates an x86-64 processor’s timestamp counter, as read by rdtsc and rdtscp.

Fields
cyclesPerSecond: u64
Implementations

impl Tsc

pub fn native_cycles_per_second() -> Option<u64>

Returns the host system’s native TSC rate, or None if it couldn’t be found.

WARNING: this is known to fail completely on some supported CPUs (particularly AMD), and can return the wrong value for others. i.e. this needs more work if we need to dependably get the host’s TSC rate. e.g. see https://github.com/shadow/shadow/issues/1519.

pub fn new(cycles_per_second: u64) -> Self

pub fn emulate_rdtsc( &self, rax: &mut u64, rdx: &mut u64, rip: &mut u64, nanos: u64, )

Updates registers to reflect the result of executing an rdtsc instruction at time nanos.

pub fn emulate_rdtscp( &self, rax: &mut u64, rdx: &mut u64, rcx: &mut u64, rip: &mut u64, nanos: u64, )

Updates registers to reflect the result of executing an rdtscp instruction at time nanos.

pub unsafe fn ip_is_rdtsc(ip: *const u8) -> bool

Whether ip points to an rdtsc instruction.

Safety

ip must be a dereferenceable pointer, pointing to the beginning of a valid x86_64 instruction.

pub unsafe fn ip_is_rdtscp(ip: *const u8) -> bool

Whether ip points to an rdtscp instruction.

Safety

ip must be a dereferenceable pointer, pointing to the beginning of a valid x86_64 instruction.